requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic
requests
aiohttp
newspaper3k
gTTS
starlette
beautifulsoup4
python-multipart